Detect a user's locale and select a language file based off of it
This attempts to provide a user with a set of words in the language
matching their locale. It first looks for a word list that matches
both the locale and territory in their locale string (e.g. fr_FR). If
it fails to find one, it looks for a wordlist appropriate for their
language (e.g. fr). If that fails, it falls back to English.