Allow user to change wordlist locale
Previously the locale could only be change by setting the LANG
environment variable. This commit adds support for mor easily changing
the locale of the wordlist to any of the supported locales shipped
with phrasebook. It can be changed either via the file menu or the new
command line flag "--locale".